# docs/swagger.yaml
basePath: /api/v1/
definitions:
  Models.GroupMessage:
    properties:
      content:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      deleted:
        type: boolean
      disliked:
        type: integer
      id:
        type: string
      liked:
        type: integer
      messageFrom:
        type: string
    type: object
  Models.Message:
    properties:
      content:
        type: string
      createdAt:
        type: string
      deleted:
        type: boolean
      disliked:
        type: integer
      id:
        type: string
      liked:
        type: integer
      messageFrom:
        type: string
      messageTo:
        type: string
      seen:
        type: boolean
    type: object
  Models.MessageBoard:
    properties:
      content:
        type: string
      createdAt:
        type: string
      deleted:
        type: boolean
      disliked:
        type: integer
      id:
        type: string
      liked:
        type: integer
      messageFrom:
        type: string
    type: object
  Models.User:
    properties:
      createdAt:
        type: string
      deleted:
        type: boolean
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      password:
        type: string
    type: object
host: localhost:8090
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a chat application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ChatApp API
  version: "1.0"
paths:
  /boardmessage:
    get:
      operationId: GetAllBoardMessages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all board messages
      tags:
      - boardmessage
    post:
      operationId: AddBoardMessage
      parameters:
      - description: Message Board data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/Models.MessageBoard'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a new board message
      tags:
      - boardmessage
  /boardmessage/{id}:
    delete:
      operationId: DeleteBoardMessageById
      parameters:
      - description: Message Board ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a board message by ID
      tags:
      - boardmessage
    get:
      operationId: GetBoardMessageById
      parameters:
      - description: Message Board ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a board message by ID
      tags:
      - boardmessage
    patch:
      operationId: UpdateBoardMessage
      parameters:
      - description: Message Board ID
        in: path
        name: id
        required: true
        type: string
      - description: Message Board data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/Models.MessageBoard'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a board message by ID
      tags:
      - boardmessage
  /boardmessage/ping:
    get:
      operationId: GetMessagesBoardPing
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ping the message board service
      tags:
      - boardmessage
  /groupmessage:
    get:
      operationId: GetAllGroupMessages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all group messages
      tags:
      - groupmessage
    post:
      operationId: AddGroupMessage
      parameters:
      - description: Group Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/Models.GroupMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a new group message
      tags:
      - groupmessage
  /groupmessage/{id}:
    delete:
      operationId: DeleteGroupMessage
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a group message by ID
      tags:
      - groupmessage
    get:
      operationId: GetGroupMessageById
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a group message by ID
      tags:
      - groupmessage
    patch:
      operationId: UpdateGroupMessage
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      - description: Group Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/Models.GroupMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a group message by ID
      tags:
      - groupmessage
  /groupmessage/ping:
    get:
      operationId: GetGroupMessagesPing
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ping the group message service
      tags:
      - groupmessage
  /message:
    get:
      operationId: GetAllMessages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all messages
      tags:
      - message
    post:
      operationId: AddMessage
      parameters:
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/Models.Message'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a new message
      tags:
      - message
  /message/{id}:
    delete:
      operationId: DeleteMessageById
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a message by ID
      tags:
      - message
    get:
      operationId: GetMessageById
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a message by ID
      tags:
      - message
    patch:
      operationId: UpdateMessage
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/Models.Message'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a message by ID
      tags:
      - message
  /message/ping:
    get:
      operationId: GetMessagesPing
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ping the message service
      tags:
      - message
  /user:
    get:
      operationId: GetAllUsers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all users
      tags:
      - user
    post:
      operationId: AddUser
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/Models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a new user
      tags:
      - user
  /user/{id}:
    delete:
      operationId: DeleteUserById
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a user by ID
      tags:
      - user
    get:
      operationId: GetUserById
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user by ID
      tags:
      - user
    patch:
      operationId: UpdateUser
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/Models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a user by ID
      tags:
      - user
  /user/ping:
    get:
      operationId: GetUserPing
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ping the server
      tags:
      - user
swagger: "2.0"